#  Add OpFunctor and replace cast, scale,clip, bce_loss and abs_grad with elementwise_no_broadcast