<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Others
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
 Add OpFunctor and replace cast, scale, full, clip, bce_loss and abs_grad with elementwise_no_broadcast
cast 当前在pten中修改不触发benchmark，现补充性能测试：

cast | case | type | old / us | new / us | speed up
-- | -- | -- | -- | -- | --
case0 | [16, 1785] | bool | 1.328 | 1.413 | 0.94
case1 | [16, 1] | int32->int64 | 1.27 | 1.239 | 1.03
case2 | [16, 1, 513, 513] | int32->float32 | 43.09 | 42.229 | 1.02
case3 | [30522, 1024] | float16->float32 | 246.57 | 236.28 | 1.04
case4 | [1] | in64->float32 | 1.276 | 1.29 | 0.99
case5 | [16, 16, 1] | float32->float16 | 1.269 | 1.312 | 0.97
case6 | [16, 16, 1024] | float16->float32 | 2.025 | 1.842 | 1.10

case0 出现性能下降，主要原因是case规模比较小，机器波动影响占比较大。